split app.go
