+++ commands.go
@@ -1,0 +1,3 @@
+package main
+
+type Command func(app *App) error
